// src/types.ts
export type PageType = 
  | 'home'
  | 'shop'
  | 'women'
  | 'men'
  | 'watches'
  | 'shoes'
  | 'accessories'
  | 'new_arrivals'
  | 'sale'
  | 'product_details'
  | 'cart'
  | 'wishlist'
  | 'checkout'
  | 'order_success'
  | 'order_tracking'
  | 'admin'
  | 'reviews'
  | 'help_support'
  | 'about';